Migrate test/e2e-encryption-perf to OpenShift Tests Extension framework
Migrate encryption performance tests to support both standard Go tests
and Ginkgo/OTE framework.

Changes:
1. Add Ginkgo test wrapper for OTE compatibility (encryption_perf.go)
2. Simplify encryption_perf_test.go to call shared test function
3. Register new test suite in main.go
4. Use library.AuthTargetGRs and library.AssertTokens from vendor

The test now works in both modes:
- Standard Go: TestPerfEncryptionTypeAESCBC(t *testing.T)
- Ginkgo/OTE: g.It(..., func(ctx) { testPerfEncryptionTypeAESCBC(ctx, g.GinkgoTB()) })
